+++ R/BandwidthSelection_KR.R
@@ -133,7 +133,8 @@ KR.bndw = function(Y, dcs_options, add_options, cf_est)
     ### Estimation of Variance Factor and Model ###
     if (isTRUE(cf_est))
     {
-      var_est = suppressWarnings(cf.estimation(R, dcs_options, add_options))
+      var_est = suppressWarnings(suppressMessages(cf.estimation(R, dcs_options,
+                                                                add_options)))
       var_coef = var_est$cf_est
       var_model = var_est$model_est
     } else {